refactor package broker.domain.exceptions
